requirements.txt
openai>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
yfinance>=0.2.35
python-dotenv>=1.0.0 